Fix session stats on 32bit OS

+++ mobile/mysterium/openvpn3/openvpn3_mock.go
@@ -81,8 +81,8 @@ type Event struct {
 
 // Statistics mock
 type Statistics struct {
-	BytesIn  int
-	BytesOut int
+	BytesIn  uint64
+	BytesOut uint64
 }
 
 // TunnelSetup mock
